Create coordinate.py
There is a robot, initially standing at the origin of the Cartesian coordinate system. The coordinates of the robot are, therefore, (0,0). Then, the robot makes N turns.

On the first turn, the robot goes 1 unit to the right.
On the second turn, the robot goes 2 units up.
On the third turn, the robot goes 3 units to the left.
On the fourth turn, the robot goes 4 units down.
On the fifth turn, the robot goes 5 units to the right.
And so on.
Your task is to determine the position of the robot after performing all the N turns.